+++ app/src/main/java/com/monke/monkeybook/presenter/ReadBookPresenterImpl.java
@@ -34,7 +34,7 @@ import com.monke.monkeybook.dao.DbHelper;
 import com.monke.monkeybook.help.ACache;
 import com.monke.monkeybook.help.BookshelfHelp;
 import com.monke.monkeybook.help.RxBusTag;
-import com.monke.monkeybook.model.BookSourceManage;
+import com.monke.monkeybook.model.BookSourceManager;
 import com.monke.monkeybook.model.ImportBookModelImpl;
 import com.monke.monkeybook.model.WebBookModelImpl;
 import com.monke.monkeybook.model.source.My716;
@@ -109,7 +109,7 @@ public class ReadBookPresenterImpl extends BasePresenterImpl<ReadBookContract.Vi
                     bookSource.setEnable(false);
                     bookSource.addGroup("禁用");
                     DbHelper.getInstance().getmDaoSession().getBookSourceBeanDao().insertOrReplace(bookSource);
-                    BookSourceManage.refreshBookSource();
+                    BookSourceManager.refreshBookSource();
                     mView.toast("已禁用" + bookSource.getBookSourceName());
                     break;
             }
