+++ app/src/main/java/com/monke/monkeybook/service/CheckSourceService.java
@@ -18,7 +18,7 @@ import com.monke.monkeybook.MApplication;
 import com.monke.monkeybook.R;
 import com.monke.monkeybook.bean.BookShelfBean;
 import com.monke.monkeybook.bean.BookSourceBean;
-import com.monke.monkeybook.model.BookSourceManage;
+import com.monke.monkeybook.model.BookSourceManager;
 import com.monke.monkeybook.model.WebBookModelImpl;
 import com.monke.monkeybook.model.analyzeRule.AnalyzeHeaders;
 import com.monke.monkeybook.model.impl.IHttpGetApi;
@@ -26,8 +26,6 @@ import com.monke.monkeybook.view.activity.BookSourceActivity;
 
 import java.net.URL;
 import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -80,7 +78,7 @@ public class CheckSourceService extends Service {
         executorService = Executors.newFixedThreadPool(threadsNum);
         scheduler = Schedulers.from(executorService);
         compositeDisposable = new CompositeDisposable();
-        bookSourceBeanList = BookSourceManage.getAllBookSource();
+        bookSourceBeanList = BookSourceManager.getAllBookSource();
         updateNotification(0);
         startCheck();
     }
@@ -199,8 +197,8 @@ public class CheckSourceService extends Service {
                             .subscribe(getObserver());
                 } catch (Exception exception) {
                     sourceBean.addGroup("失效");
-                    BookSourceManage.addBookSource(sourceBean);
-                    BookSourceManage.refreshBookSource();
+                    BookSourceManager.addBookSource(sourceBean);
+                    BookSourceManager.refreshBookSource();
                     nextCheck();
                 }
             } else {
@@ -214,8 +212,8 @@ public class CheckSourceService extends Service {
                             .subscribe(getObserver());
                 } catch (Exception e) {
                     sourceBean.addGroup("失效");
-                    BookSourceManage.addBookSource(sourceBean);
-                    BookSourceManage.refreshBookSource();
+                    BookSourceManager.addBookSource(sourceBean);
+                    BookSourceManager.refreshBookSource();
                     nextCheck();
                 }
             }
@@ -239,8 +237,8 @@ public class CheckSourceService extends Service {
                 public void onNext(Object value) {
                     if (sourceBean.containsGroup("失效")) {
                         sourceBean.removeGroup("失效");
-                        BookSourceManage.addBookSource(sourceBean);
-                        BookSourceManage.refreshBookSource();
+                        BookSourceManager.addBookSource(sourceBean);
+                        BookSourceManager.refreshBookSource();
                     }
                     nextCheck();
                 }
@@ -249,8 +247,8 @@ public class CheckSourceService extends Service {
                 public void onError(Throwable e) {
                     sourceBean.addGroup("失效");
                     sourceBean.setSerialNumber(10000 + checkIndex);
-                    BookSourceManage.addBookSource(sourceBean);
-                    BookSourceManage.refreshBookSource();
+                    BookSourceManager.addBookSource(sourceBean);
+                    BookSourceManager.refreshBookSource();
                     nextCheck();
                 }
