+++ app/src/main/java/com/monke/monkeybook/view/activity/SourceEditActivity.java
@@ -459,8 +459,7 @@ public class SourceEditActivity extends MBaseActivity<SourceEditContract.Present
                 new AlertDialog.Builder(this)
                         .setTitle(getString(R.string.exit))
                         .setMessage(getString(R.string.exit_no_save))
-                        .setPositiveButton("是", (DialogInterface dialogInterface, int which) -> {
-                        })
+                        .setPositiveButton("是", (DialogInterface dialogInterface, int which) -> { })
                         .setNegativeButton("否", (DialogInterface dialogInterface, int which) -> finish())
                         .show();
                 return true;
